+++ changelog.md
@@ -2,6 +2,8 @@
 
 - For some reason, the Discord Invite got removed for no apparently reason lol
 - Removed part of an anti-cheat
+- Added Amethyst Reward Animation
+- Internal changes
 
 ## v0.1.0-beta.1